+++ lib/flutter_recorder.dart
@@ -3,8 +3,14 @@ library flutter_recorder;
 
 export 'src/audio_data_container.dart';
 export 'src/enums.dart'
-    show CaptureDevice, CaptureErrors, PCMFormat, RecorderChannels;
+    show
+        AndroidInputPreset,
+        CaptureDevice,
+        CaptureErrors,
+        PCMFormat,
+        RecorderChannels;
 export 'src/filters/autogain.dart';
 export 'src/filters/echo_cancellation.dart';
-export 'src/filters/filters.dart' show RecorderFilterType;
+export 'src/filters/filters.dart'
+    show FilterMetric, FilterParam, RecorderFilterType;
 export 'src/flutter_recorder.dart';